+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 v1.25 (??? 2022)
     * Allow negative dates in `date2date`.
     * Support of microseconds in `date2date`.
+    * Assure that `input2array` gives no 0d-array.
 
 v1.24 (Jun 2022)
     * Removed documentation from `Read the Docs` and moved to Github Pages: